# frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Resume Builder API specification",
    "description": "API for Resume Builder",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8080/api/v1", "description": "Local server" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/sections/{id}": {
      "get": {
        "tags": ["Resume Sections"],
        "operationId": "getSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SectionResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Resume Sections"],
        "operationId": "updateSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SectionRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/layouts": {
      "post": {
        "tags": ["Layout"],
        "operationId": "createLayout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LayoutRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/layouts/{id}": {
      "get": {
        "tags": ["Layout"],
        "operationId": "getLayout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LayoutResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Layout"],
        "operationId": "updateLayout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LayoutRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/layouts/{id}/delete": {
      "post": {
        "tags": ["Layout"],
        "operationId": "deleteLayout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/latexmethods": {
      "post": {
        "tags": ["LatexMethod"],
        "operationId": "createLatexMethod",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LatexMethodRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/latexmethods/{id}": {
      "get": {
        "tags": ["LatexMethod"],
        "operationId": "getLatexMethod",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LatexMethodResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["LatexMethod"],
        "operationId": "updateLatexMethod",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LatexMethodRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/latexmethods/{id}/delete": {
      "post": {
        "tags": ["LatexMethod"],
        "operationId": "deleteLatexMethod",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/columnsections": {
      "post": {
        "tags": ["ColumnSection"],
        "operationId": "createColumnSection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ColumnSectionRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/columnsections/{id}": {
      "get": {
        "tags": ["ColumnSection"],
        "operationId": "getColumnSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnSectionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ColumnSection"],
        "operationId": "updateColumnSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ColumnSectionRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/columnsections/{id}/delete": {
      "post": {
        "tags": ["ColumnSection"],
        "operationId": "deleteColumnSection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/columns": {
      "post": {
        "tags": ["Column"],
        "operationId": "createColumn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ColumnRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "format": "int64" }
              }
            }
          }
        }
      }
    },
    "/columns/{id}": {
      "get": {
        "tags": ["Column"],
        "operationId": "getColumn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ColumnResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Column"],
        "operationId": "updateColumn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ColumnRequest" }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/columns/{id}/delete": {
      "post": {
        "tags": ["Column"],
        "operationId": "deleteColumn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/layouts/{id}/pdf": {
      "get": {
        "tags": ["Layout"],
        "operationId": "generateLatexPdf",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          }
        }
      }
    },
    "/layouts/{id}/methods": {
      "get": {
        "tags": ["Layout"],
        "operationId": "getLatexMethods",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LatexMethodResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SectionItemRequest": {
        "required": ["data", "latexMethodId", "type"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "sectionId": { "type": "integer", "format": "int64" },
          "latexMethodId": { "type": "integer", "format": "int64" },
          "type": { "type": "string" },
          "itemOrder": { "type": "integer", "format": "int32" },
          "alignment": {
            "type": "string",
            "enum": ["LEFT", "CENTER", "RIGHT", "JUSTIFIED"]
          },
          "data": {
            "type": "object",
            "additionalProperties": { "type": "object" }
          }
        }
      },
      "SectionRequest": {
        "required": ["showTitle", "title"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "resumeId": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "showTitle": { "type": "boolean" },
          "sectionItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionItemRequest" }
          }
        }
      },
...
      "ColorScheme": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "primaryColor": { "pattern": "^#[0-9A-Fa-f]{6}$", "type": "string" },
          "secondaryColor": {
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "type": "string"
          },
          "accent": { "pattern": "^#[0-9A-Fa-f]{6}$", "type": "string" },
          "darkBg": { "pattern": "^#[0-9A-Fa-f]{6}$", "type": "string" },
          "lightBg": { "pattern": "^#[0-9A-Fa-f]{6}$", "type": "string" },
          "darkText": { "pattern": "^#[0-9A-Fa-f]{6}$", "type": "string" },
          "lightText": { "pattern": "^#[0-9A-Fa-f]{6}$", "type": "string" }
        }
      },
      "ColumnRequest": {
        "required": ["sectionMappings"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "layoutId": { "type": "integer", "format": "int64" },
          "columnNumber": {
            "maximum": 2,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sectionMappings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ColumnSectionRequest" }
          },
          "backgroundColor": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "ACCENT",
              "DARK_BG",
              "LIGHT_BG",
              "DARK_TEXT",
              "LIGHT_TEXT"
            ]
          },
          "textColor": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "ACCENT",
              "DARK_BG",
              "LIGHT_BG",
              "DARK_TEXT",
              "LIGHT_TEXT"
            ]
          },
          "borderColor": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "ACCENT",
              "DARK_BG",
              "LIGHT_BG",
              "DARK_TEXT",
              "LIGHT_TEXT"
            ]
          },
          "paddingLeft": { "minimum": 0, "type": "number", "format": "double" },
          "paddingRight": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "paddingTop": { "minimum": 0, "type": "number", "format": "double" },
          "paddingBottom": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "borderLeft": { "minimum": 0, "type": "number", "format": "double" },
          "borderRight": { "minimum": 0, "type": "number", "format": "double" },
          "borderTop": { "minimum": 0, "type": "number", "format": "double" },
          "borderBottom": { "minimum": 0, "type": "number", "format": "double" }
        }
      },
      "ColumnSectionRequest": {
        "required": ["columnId", "sectionId"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "columnId": { "type": "integer", "format": "int64" },
          "sectionId": { "type": "integer", "format": "int64" },
          "sectionOrder": { "type": "integer", "format": "int32" },
          "itemsep": { "type": "number", "format": "double" },
          "endsep": { "type": "number", "format": "double" }
        }
      },
      "LatexMethodRequest": {
        "required": ["method", "name", "type"],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "layoutId": { "type": "integer", "format": "int64" },
          "type": {
            "type": "string",
            "enum": [
              "TITLE",
              "WORK_EXPERIENCE",
              "TEXTBOX",
              "SKILL",
              "EDUCATION",
              "SKILL_BOXES",
              "PICTURE",
              "CONTACT"
            ]
          },
          "name": { "type": "string" },
          "method": { "type": "string" }
        }
      },
      "LayoutRequest": {
        "required": [
          "colorScheme",
          "columnSeparator",
          "latexMethods",
          "numberOfColumns",
          "resumeId"
        ],
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "resumeId": { "type": "integer", "format": "int64" },
          "pageSize": { "type": "string", "enum": ["A4", "LETTER"] },
          "columns": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ColumnRequest" }
          },
          "numberOfColumns": {
            "maximum": 2,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "columnSeparator": {
            "maximum": 1,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "colorScheme": { "$ref": "#/components/schemas/ColorScheme" },
          "latexMethods": {
            "uniqueItems": true,
            "type": "array",
            "items": { "$ref": "#/components/schemas/LatexMethodRequest" }
          },
          "sectionTitleMethod": { "type": "string" }
        }
      },
      "LatexMethodResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "type": {
            "type": "string",
            "enum": [
              "TITLE",
              "WORK_EXPERIENCE",
              "TEXTBOX",
              "SKILL",
              "EDUCATION",
              "SKILL_BOXES",
              "PICTURE",
              "CONTACT"
            ]
          },
          "name": { "type": "string" },
          "method": { "type": "string" }
        }
      },
      "SectionItemResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "latexMethod": { "$ref": "#/components/schemas/LatexMethodResponse" },
          "type": { "type": "string" },
          "itemOrder": { "type": "integer", "format": "int32" },
          "alignment": {
            "type": "string",
            "enum": ["LEFT", "CENTER", "RIGHT", "JUSTIFIED"]
          },
          "data": {
            "type": "object",
            "additionalProperties": { "type": "object" }
          }
        }
      },
      "SectionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "showTitle": { "type": "boolean" },
          "sectionItems": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SectionItemResponse" }
          }
        }
      },
...
      "ColumnResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "columnNumber": { "type": "integer", "format": "int32" },
          "sectionMappings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ColumnSectionResponse" }
          },
          "backgroundColor": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "ACCENT",
              "DARK_BG",
              "LIGHT_BG",
              "DARK_TEXT",
              "LIGHT_TEXT"
            ]
          },
          "textColor": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "ACCENT",
              "DARK_BG",
              "LIGHT_BG",
              "DARK_TEXT",
              "LIGHT_TEXT"
            ]
          },
          "borderColor": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY",
              "ACCENT",
              "DARK_BG",
              "LIGHT_BG",
              "DARK_TEXT",
              "LIGHT_TEXT"
            ]
          },
          "paddingLeft": { "type": "number", "format": "double" },
          "paddingRight": { "type": "number", "format": "double" },
          "paddingTop": { "type": "number", "format": "double" },
          "paddingBottom": { "type": "number", "format": "double" },
          "borderLeft": { "type": "number", "format": "double" },
          "borderRight": { "type": "number", "format": "double" },
          "borderTop": { "type": "number", "format": "double" },
          "borderBottom": { "type": "number", "format": "double" }
        }
      },
      "ColumnSectionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "section": { "$ref": "#/components/schemas/SectionResponse" },
          "sectionOrder": { "type": "integer", "format": "int32" },
          "itemsep": { "type": "number", "format": "double" },
          "endsep": { "type": "number", "format": "double" }
        }
      },
      "LayoutResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "pageSize": { "type": "string", "enum": ["A4", "LETTER"] },
          "columns": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ColumnResponse" }
          },
          "numberOfColumns": { "type": "integer", "format": "int32" },
          "columnSeparator": { "type": "number", "format": "double" },
          "colorScheme": { "$ref": "#/components/schemas/ColorScheme" },
          "latexMethods": {
            "uniqueItems": true,
            "type": "array",
            "items": { "$ref": "#/components/schemas/LatexMethodResponse" }
          },
          "sectionTitleMethod": { "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}